회원정보 수정 후 기존 토큰 복사해서 nickName claim 업데이트 후 토큰 다시 발행/응답

+++ src/main/java/com/prj2spring20240521/controller/board/BoardController.java
@@ -3,6 +3,7 @@ package com.prj2spring20240521.controller.board;
 import com.prj2spring20240521.domain.board.Board;
 import com.prj2spring20240521.service.board.BoardService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
@@ -59,13 +60,23 @@ public class BoardController {
     }
 
     @DeleteMapping("{id}")
+    @PreAuthorize("isAuthenticated()")
     // @PathVariable : URL 경로 변수의 값을 메서드 파라미터로 바인딩해준다.
-    public void delete(@PathVariable Integer id) {
-        service.remove(id);
+    public ResponseEntity<Object> delete(@PathVariable Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Board board) {
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity edit(@RequestBody Board board, Authentication authentication) {
+        if (!service.hasAccess(board.getId(), authentication)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         if (service.validate(board)) {
             service.edit(board);
             return ResponseEntity.ok().build();
